reverse/tcp-madness/src/requirements.txt
schedule
pycrypto
